Добавляет исправления по адаптивным сеткам
